Replaced points formula with FINA formula Added command to recalculate points Points are now stored as integer

+++ rankings/models.py
@@ -316,7 +316,7 @@ class IndividualResult(models.Model):
                              related_name='individual_results',
                              related_query_name='individual_results')
     time = models.DurationField(blank=True, null=True)
-    points = models.FloatField(default=0)
+    points = models.IntegerField(default=0)
     original_line = models.CharField(max_length=200, null=True, default=None)
     round = models.IntegerField(default=0)
     disqualified = models.BooleanField(default=False)
